server/hu-bmap-ontology-api-client/hu_bmap_ontology_api_client/models/concept_detail.py
from typing import Any, Dict, List, Type, TypeVar, Union

import attr

from ..types import UNSET, Unset

T = TypeVar("T", bound="ConceptDetail")


@attr.s(auto_attribs=True)
class ConceptDetail:
    """ """

    concept: Union[Unset, str] = UNSET
    perfterm: Union[Unset, str] = UNSET
    additional_properties: Dict[str, Any] = attr.ib(init=False, factory=dict)

    def to_dict(self) -> Dict[str, Any]:
        concept = self.concept
        perfterm = self.perfterm

        field_dict: Dict[str, Any] = {}
        field_dict.update(self.additional_properties)
        field_dict.update({})
        if concept is not UNSET:
            field_dict["concept"] = concept
        if perfterm is not UNSET:
            field_dict["perfterm"] = perfterm

        return field_dict

    @classmethod
    def from_dict(cls: Type[T], src_dict: Dict[str, Any]) -> T:
        d = src_dict.copy()
        concept = d.pop("concept", UNSET)

        perfterm = d.pop("perfterm", UNSET)

        concept_detail = cls(
            concept=concept,
            perfterm=perfterm,
        )

        concept_detail.additional_properties = d
        return concept_detail

    @property
    def additional_keys(self) -> List[str]:
        return list(self.additional_properties.keys())

    def __getitem__(self, key: str) -> Any:
        return self.additional_properties[key]

    def __setitem__(self, key: str, value: Any) -> None:
        self.additional_properties[key] = value

    def __delitem__(self, key: str) -> None:
        del self.additional_properties[key]

    def __contains__(self, key: str) -> bool:
        return key in self.additional_properties
